+++ Changelog.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented here.
 - Fixed Haskell tester installation using ghcup to install stack system-wide (#688)
 - Updated tester schema generation to use msgspec datatypes (#689)
 - Specify dockerfile frontend version (#691)
+- Add a new Javascript (Jest) tester to support Javascript autotesting (#698)
 - Add remote URL whitelist for AI tester to restrict allowed endpoints (#693)
 - Increased default settings job timeout from 600s to 1200s (#707)